Renamed StudyInfoHelpers to TutkinnonOsaHelpers
