+++ shared/components/TukinnonOsa/Objectives.tsx
@@ -3,7 +3,7 @@ import styled from "../../styled"
 import { FormattedMessage, intlShape } from "react-intl"
 import { Expand } from "./Expand"
 import { IconContainer } from "./IconContainer"
-import { ToggleableItems } from "./StudyInfoHelpers"
+import { ToggleableItems } from "./TutkinnonOsaHelpers"
 import { Collapse } from "./Collapse"
 import { HorizontalLine } from "../HorizontalLine"
 import { InfoTextContainer } from "./Shared"
