+++ shared/components/TukinnonOsa/TutkinnonOsaHelpers.ts
