+++ shared/components/TutkinnonOsa.tsx
@@ -16,7 +16,7 @@ import { FormattedMessage } from "react-intl"
 import { navigate } from "@reach/router"
 import { stringifyShareParams } from "utils/shareParams"
 import { AppContext } from "components/AppContext"
-import { ToggleableItems } from "./TukinnonOsa/StudyInfoHelpers"
+import { ToggleableItems } from "./TukinnonOsa/TutkinnonOsaHelpers"
 import { Objectives } from "./TukinnonOsa/Objectives"
 import { Details } from "./TukinnonOsa/Details"
 import { OneRowTable } from "./TukinnonOsa/Shared"
